strands/interview-prep-agent/requirements.txt
strands-agents[openai]
strands-agents-tools
mcp
pydantic>=2.0
python-dotenv
colorama
langsmith[otel]